Ununmake
Removed unmake methods is favour of storing copies and reversable flags.

TreeSearch now stores a copy of the board cards, calls test_move which sets card.played flag to prevent further play. p_turn is passed around to keep track of which player's turn and the difference to player's score is kept track of in a dict.